Separated out ModuleInterfacce and NodeFactory class definitions to their repective headers. Much cleaner.
