+++ Source/GenericGraphEditor/Private/GenericGraphEditor.cpp
@@ -1,50 +1,11 @@
+#include "GenericGraphEditor.h"
+#include "GenericGraphNodeFactory.h"
 #include "AssetTypeActions_GenericGraph.h"
-#include "GenericGraphEditorPCH.h"
-#include "EdGraphUtilities.h"
-#include "GenericGraphAssetEditor/EdNode_GenericGraphNode.h"
-#include "GenericGraphAssetEditor/EdNode_GenericGraphEdge.h"
-#include "GenericGraphAssetEditor/SEdNode_GenericGraphNode.h"
-#include "GenericGraphAssetEditor/SEdNode_GenericGraphEdge.h"
 #include "GenericGraphAssetEditor/GenericGraphEditorStyle.h"
 
 DEFINE_LOG_CATEGORY(GenericGraphEditor)
 
 #define LOCTEXT_NAMESPACE "Editor_GenericGraph"
-
-class FGraphPanelNodeFactory_GenericGraph : public FGraphPanelNodeFactory
-{
-	virtual TSharedPtr<class SGraphNode> CreateNode(UEdGraphNode* Node) const override
-	{
-		if (UEdNode_GenericGraphNode* EdNode_GraphNode = Cast<UEdNode_GenericGraphNode>(Node))
-		{
-			return SNew(SEdNode_GenericGraphNode, EdNode_GraphNode);
-		}
-		else if (UEdNode_GenericGraphEdge* EdNode_Edge = Cast<UEdNode_GenericGraphEdge>(Node))
-		{
-			return SNew(SEdNode_GenericGraphEdge, EdNode_Edge);
-		}
-		return nullptr;
-	}
-};
-
-TSharedPtr<FGraphPanelNodeFactory> GraphPanelNodeFactory_GenericGraph;
-
-class FGenericGraphEditor : public IGenericGraphEditor
-{
-	/** IModuleInterface implementation */
-	virtual void StartupModule() override;
-	virtual void ShutdownModule() override;
-
-private:
-	void RegisterAssetTypeAction(IAssetTools& AssetTools, TSharedRef<IAssetTypeActions> Action);
-
-private:
-	TArray< TSharedPtr<IAssetTypeActions> > CreatedAssetTypeActions;
-
-	EAssetTypeCategories::Type GenericGraphAssetCategoryBit;
-};
-
-IMPLEMENT_MODULE( FGenericGraphEditor, GenericGraphEditor )
 
 void FGenericGraphEditor::StartupModule()
 {
