+++ Source/GenericGraphEditor/Public/GenericGraphNodeFactory.h
@@ -1,0 +1,8 @@
+#pragma once
+#include <EdGraphUtilities.h>
+#include <EdGraph/EdGraphNode.h>
+
+class FGraphPanelNodeFactory_GenericGraph : public FGraphPanelNodeFactory
+{
+	virtual TSharedPtr<class SGraphNode> CreateNode(UEdGraphNode* Node) const override;
+};
